Refactor navigation menu components
