+++ src/app/demo/[name]/ui/navigation-menu-secondary.tsx
@@ -1,0 +1,117 @@
+import { NavigationMenu, NavigationMenuLink, NavigationMenuContent, NavigationMenuItem, NavigationMenuList, navigationMenuTriggerStyle, NavigationMenuTrigger } from "@/components/ui/navigation-menu";
+import { Icon } from "@/lib/icon";
+import { mdiCheckCircleOutline, mdiCircleOutline, mdiHelpCircleOutline } from "@mdi/js";
+
+export default function NavigationMenuSecondary() {
+    return (
+        <div className="pb-50">
+            <NavigationMenu viewport={false}>
+                <NavigationMenuList>
+                    <NavigationMenuItem>
+                        <NavigationMenuLink
+                            asChild
+                            className={navigationMenuTriggerStyle()}
+                        >
+                            <Link href="#">Documentation</Link>
+                        </NavigationMenuLink>
+                    </NavigationMenuItem>
+
+                    <NavigationMenuItem>
+                        <NavigationMenuTrigger>List</NavigationMenuTrigger>
+                        <NavigationMenuContent>
+                            <ul className="grid w-[300px] gap-4">
+                                <li>
+                                    <NavigationMenuLink asChild>
+                                        <Link href="#">
+                                            <div className="font-medium">Components</div>
+                                            <div className="text-muted-foreground">
+                                                Browse all components in the library.
+                                            </div>
+                                        </Link>
+                                    </NavigationMenuLink>
+                                    <NavigationMenuLink asChild>
+                                        <Link href="#">
+                                            <div className="font-medium">Documentation</div>
+                                            <div className="text-muted-foreground">
+                                                Learn how to use the library.
+                                            </div>
+                                        </Link>
+                                    </NavigationMenuLink>
+                                    <NavigationMenuLink asChild>
+                                        <Link href="#">
+                                            <div className="font-medium">Blog</div>
+                                            <div className="text-muted-foreground">
+                                                Read our latest blog posts.
+                                            </div>
+                                        </Link>
+                                    </NavigationMenuLink>
+                                </li>
+                            </ul>
+                        </NavigationMenuContent>
+                    </NavigationMenuItem>
+
+                    <NavigationMenuItem>
+                        <NavigationMenuTrigger>Simple List</NavigationMenuTrigger>
+                        <NavigationMenuContent>
+                            <ul className="grid w-[200px] gap-4">
+                                <li>
+                                    <NavigationMenuLink asChild>
+                                        <Link href="#">Components</Link>
+                                    </NavigationMenuLink>
+                                    <NavigationMenuLink asChild>
+                                        <Link href="#">Documentation</Link>
+                                    </NavigationMenuLink>
+                                    <NavigationMenuLink asChild>
+                                        <Link href="#">Blocks</Link>
+                                    </NavigationMenuLink>
+                                </li>
+                            </ul>
+                        </NavigationMenuContent>
+                    </NavigationMenuItem>
+
+                    <NavigationMenuItem>
+                        <NavigationMenuTrigger>With Icon</NavigationMenuTrigger>
+                        <NavigationMenuContent>
+                            <ul className="grid w-[200px] gap-4">
+                                <li>
+                                    <NavigationMenuLink asChild>
+                                        <Link href="#" className="flex-row items-center gap-2">
+                                            <Icon path={mdiHelpCircleOutline} />
+                                            Backlog
+                                        </Link>
+                                    </NavigationMenuLink>
+                                    <NavigationMenuLink asChild>
+                                        <Link href="#" className="flex-row items-center gap-2">
+                                            <Icon path={mdiCircleOutline} />
+                                            To Do
+                                        </Link>
+                                    </NavigationMenuLink>
+                                    <NavigationMenuLink asChild>
+                                        <Link href="#" className="flex-row items-center gap-2">
+                                            <Icon path={mdiCheckCircleOutline} />
+                                            Done
+                                        </Link>
+                                    </NavigationMenuLink>
+                                </li>
+                            </ul>
+                        </NavigationMenuContent>
+                    </NavigationMenuItem>
+                </NavigationMenuList>
+            </NavigationMenu>
+        </div>
+    )
+}
+
+// Framework-agnostic Link component - simple anchor tag for demo purposes
+function Link({
+    href,
+    children,
+    className,
+    ...props
+}: React.AnchorHTMLAttributes<HTMLAnchorElement> & { href: string }) {
+    return (
+        <a href={href} className={className} {...props}>
+            {children}
+        </a>
+    );
+}
